requirements.txt
coqpit==0.0.17
ffmpeg-python==0.2.0
ffmpy==0.3.1
Flask @ file:///croot/flask_1702980023242/work
numpy==1.24.4
openai==1.10.0
openai-whisper==20231117
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.11
pybind11==2.11.1
pycountry==23.12.11
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments @ file:///croot/pygments_1684279966437/work
pynndescent==0.5.11
pynvml @ file:///home/conda/feedstock_root/build_artifacts/pynvml_1676513396402/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pypinyin==0.50.0
pyproject-api==1.6.1
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
pysbd==0.3.4
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytablewriter==1.2.0
python-crfsuite==0.9.10
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1686601365461/work
referencing==0.33.0
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
scikit-learn==1.4.0
scipy==1.11.4
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition @ file:///home/conda/feedstock_root/build_artifacts/speechrecognition_1701879627473/work
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.1
toml==0.10.2
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
toolz==0.12.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado @ file:///croot/tornado_1696936946304/work
tox==4.12.1
tqdm @ file:///croot/tqdm_1679561862951/work
tqdm-multiprocess==0.0.11
trainer==0.0.36
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.37.1
triton==2.1.0
TTS==0.22.0
typepy==1.3.2
typer==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
ucimlrepo==0.0.3
umap-learn==0.5.5
Unidecode==1.3.8
urllib3 @ file:///croot/urllib3_1698257533958/work
uvicorn==0.27.0.post1
virtualenv==20.25.0
Werkzeug @ file:///croot/werkzeug_1679489717957/work
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0